% Generated by build script
\name{ia_metadata}
\alias{ia_metadata}
\title{Get metadata for an Internet Archive item}
\usage{
ia_metadata()
}
\arguments{
\item{identifier}{The item identifier (e.g., "gov.uscourts.dcd.123456")}
}
\value{tibble: field, value}
\description{Get metadata for an Internet Archive item}
